+++ smawg/cli.py
@@ -213,11 +213,7 @@ class Client:
         print(tabulate(rows, headers, stralign="center", numalign="center"))
 
     def _command_show_regions(self, args: list[str]) -> None:
-        if len(args) == 0:
-            raise InvalidCommand("You need to provide a player index")
-        if len(args) > 1:
-            raise InvalidCommand("'show-regions' expects only one argument")
-        i = self._parse_int(args[0])
+        i = self._parse_ints(args, n=1)[0]
         if not 0 <= i < len(self.game.players):
             msg = f"<player> must be between 0 and {len(self.game.players)}"
             raise ValueError(msg)
@@ -230,29 +226,23 @@ class Client:
         print(tabulate(rows, headers, stralign="center", numalign="center"))
 
     def _command_combo(self, args: list[str]) -> None:
-        if len(args) == 0:
-            raise InvalidCommand("You need to provide a combo index")
-        if len(args) > 1:
-            raise InvalidCommand("'combo' expects only one argument")
-        i = self._parse_int(args[0])
+        i = self._parse_ints(args, n=1)[0]
         self.game.select_combo(i)
 
     def _command_conquer(self, args: list[str]) -> None:
-        if len(args) == 0:
-            raise InvalidCommand("You need to provide a region index")
-        if len(args) > 1:
-            raise InvalidCommand("'conquer' expects only one argument")
-        i = self._parse_int(args[0])
+        i = self._parse_ints(args, n=1)[0]
         self.game.conquer(i)
 
     def _command_deploy(self, args: list[str]) -> None:
-        if len(args) < 2:
-            msg = "You need to provide a number of tokens and a region index"
-            raise InvalidCommand(msg)
-        if len(args) > 2:
-            raise InvalidCommand("'deploy' expects only 2 arguments")
-        n, region = [self._parse_int(a) for a in args]
+        n, region = self._parse_ints(args, n=2)
         self.game.deploy(n, region)
+
+    def _parse_ints(self, args: list[str], *, n: int) -> list[int]:
+        """Parse `args` as a list of `n` integers."""
+        if len(args) != n:
+            msg = f"expected {n} argument(s), but got {len(args)}"
+            raise InvalidCommand(msg)
+        return [self._parse_int(a) for a in args]
 
     def _parse_int(self, s: str) -> int:
         """Parse an integer or raise `ValueError` with a frendly message."""