+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-\-
+### Changed
+- `cli` error messages when given wrong number of arguments.
 
 ## [0.4.0] - 2022-04-12
 ### Added