+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+\-
+
+## [0.5.0] - 2022-04-18
 ### Added
 - Reinforcements dice mechanic.
 - Related `Game.conquer()` argument: `use_dice`.